feat: added creating a property to state store
