+++ frontend/src/components/organism/Admin/Panel.jsx
@@ -3,8 +3,10 @@ import { useNavigate } from "react-router-dom";
 import { button } from "../../atom";
 
 import styles from "../../../styles/components/organism/Admin.module.css";
+import useAppContext from "../../../hooks/useAppContext";
 
 const Panel = () => {
+  const { dispatch } = useAppContext();
   const navigate = useNavigate();
 
   return (
@@ -15,7 +17,12 @@ const Panel = () => {
       </div>
       <div className={styles.right}>
         {button(
-          () => console.log("You clicked me"),
+          () => {
+            // load form for modal to display
+            dispatch({ type: "WHICH FORM TO USE", payload: "CREATE PROPERTY" });
+            // show modal on button click
+            dispatch({ type: "CHANGE MODAL STATE", payload: true });
+          },
           "Add New Property",
           "success",
         )}
