+++ frontend/src/components/organism/Forms/auth/Logout.jsx
@@ -4,14 +4,14 @@ import Form from "react-bootstrap/Form";
 import { useForm } from "react-hook-form";
 
 // atomic design
-import { button } from "../../atom";
+import { button } from "../../../atom";
 
 // utils and custom hooks
-import useAppContext from "../../../hooks/useAppContext";
-import { onSubmit } from "../../../utils";
+import useAppContext from "../../../../hooks/useAppContext";
+import { onSubmit } from "../../../../utils";
 
 // styling
-import styles from "../../../styles/components/organism/Forms.module.css";
+import styles from "../../../../styles/components/organism/Forms.module.css";
 
 /**
  * Logout form currently being used for displaying in the Modal
