+++ frontend/src/components/organism/Forms/auth/Registration.jsx
@@ -4,50 +4,55 @@ import { useEffect } from "react";
 import { useForm } from "react-hook-form";
 
 // atomic design
-import MapToForm from "../../molecule/MapToForm";
-import { button, spinner } from "../../atom";
+import MapToForm from "../../../molecule/MapToForm";
+import { button, spinner } from "../../../atom";
 
 // utils and custom hooks
-import useAppContext from "../../../hooks/useAppContext";
-import { onSubmit } from "../../../utils";
+import useAppContext from "../../../../hooks/useAppContext";
+import { onSubmit } from "../../../../utils";
 
 // styling
-import styles from "../../../styles/components/organism/Forms.module.css";
+import styles from "../../../../styles/components/organism/Forms.module.css";
 
 /**
- * Change a Password form this is being displayed inside the modal
+ * Registration Form currently being used for displaying in the modal
  * @returns {JSX.Element}
  * @constructor
  */
-const ChangePassword = () => {
+const Registration = () => {
   // state store
   const { dispatch, forms } = useAppContext();
-  const { loading, err } = forms;
+  const { loading } = forms;
 
-  // React hook form functions
+  // React hook forms functions
   const {
     register,
     handleSubmit,
+    watch,
     formState: { errors },
   } = useForm();
 
   useEffect(() => {
     // change the header of the modal
-    dispatch({ type: "CHANGE HEADER", payload: "Change Password Form" });
+    dispatch({ type: "CHANGE HEADER", payload: "Registration Form" });
+    // change the submit button text on modal
+    dispatch({ type: "CHANGE BTN", payload: "Registration" });
+    // change the url needed for the submitted button
+    dispatch({
+      type: "FORM SUBMIT URL",
+      payload: "auth/registration/",
+    });
   }, [dispatch]);
 
-  // load login form
+  // Login user form will load
   const handleLogin = () =>
     dispatch({ type: "WHICH FORM TO USE", payload: "" });
 
-  // load register form
-  const handleRegister = () =>
-    dispatch({ type: "WHICH FORM TO USE", payload: "REGISTRATION" });
-
   // password checking
+  const passwordMatch = watch("password1");
   const commonPasswords = ["123456", "password", "12345678", "qwerty"];
 
-  // array for MapToForm molecule
+  // array used for MapToForm molecule
   const arr = [
     {
       id: 1,
@@ -58,14 +63,29 @@ const ChangePassword = () => {
     },
     {
       id: 2,
-      name: "old_password",
+      name: "password1",
       type: "password",
       placeholder: "Your Password",
-      formValidation: { required: "This field is required" },
+      formValidation: {
+        required: "This field is required",
+        minLength: {
+          value: 8,
+          message:
+            "This password is too short. It must contain at least 8 characters",
+        },
+        validate: {
+          // validation to check if password is not too common
+          // and for checking if password is not all numeric
+          common: (value) =>
+            !commonPasswords.includes(value) || "This password is too common",
+          numeric: (value) =>
+            !/^\d+$/.test(value) || "This password is entirely numeric",
+        },
+      },
     },
     {
       id: 3,
-      name: "new_password",
+      name: "password2",
       type: "password",
       placeholder: "Re-type Password",
       formValidation: {
@@ -76,7 +96,8 @@ const ChangePassword = () => {
             "This password is too short. It must contain at least 8 characters",
         },
         validate: {
-          // Validation to check that password is not too common or all numeric
+          // validates if passwords match, common passwords and password is not all numeric
+          match: (value) => value === passwordMatch || "Passwords do not match",
           common: (value) =>
             !commonPasswords.includes(value) || "This password is too common",
           numeric: (value) =>
@@ -89,19 +110,15 @@ const ChangePassword = () => {
   return (
     <Form
       onSubmit={handleSubmit((data) =>
-        onSubmit(data, "auth/change_password/", dispatch),
+        onSubmit(data, "auth/registration/", dispatch),
       )}
     >
-      {/*Display spinner when waiting for server response*/}
       {loading ? spinner() : null}
 
-      {/*Buttons to link back to Register and Login Form*/}
       <div className={styles.btn}>
-        {button(handleRegister, "Click here to Register?", "secondary")}
         {button(handleLogin, "Login Here?", "primary")}
       </div>
 
-      {/*displaying of Form Control input elements and error messages*/}
       {arr.map(({ id, name, type, placeholder, formValidation }) => (
         <MapToForm
           key={id}
@@ -114,20 +131,17 @@ const ChangePassword = () => {
         />
       ))}
 
-      {/*Cancel and Submit buttons*/}
+      {/*close modal or login buttons*/}
       <div className={styles.btn}>
         {button(
           () => dispatch({ type: "CHANGE MODAL STATE", payload: false }),
           "Cancel",
           "secondary",
         )}
-        {button("submit", "Change Password", "warning")}
+        {button("submit", "Register", "warning")}
       </div>
-
-      {/*display an error message only if they occur*/}
-      {err["message"] && <p className="text-danger">{err["message"]}</p>}
     </Form>
   );
 };
 
-export default ChangePassword;
+export default Registration;
